requirements.txt
langchain_openai
langchain_community
langgraph
pydantic
fastapi
uvicorn
python-dotenv